VisualRobot_wzh：新增猫狗二分类人物测试图像+模型，main.cpp中新增示例程序。项目无报错、警告，板上运行无误。

+++ VisualRobot/featureDetect.cpp
@@ -65,7 +65,7 @@ std::vector<cv::DMatch> FeatureDetector::filterMatches(
                 }
             }
 
-            if (ransacMatches.size() >= params.minInliers) {
+            if (ransacMatches.size() >= static_cast<size_t>(params.minInliers)) {
                 points1 = inlierPoints1;
                 points2 = inlierPoints2;
                 return ransacMatches;
